Renaming to NativeFunction and SemanticFunction
